[Refactor] Remove unused packages. Refactor banners for each page
