+++ package.json
@@ -3,8 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
     "@nivo/core": "^0.74.0",
     "@nivo/pie": "^0.74.0",
     "@reach/dialog": "^0.16.2",
@@ -19,19 +17,14 @@
     "firebase": "^9.1.3",
     "firebase-tools": "^9.20.0",
     "google-map-react": "^2.1.10",
-    "lodash": "^4.17.21",
-    "moment": "^2.29.1",
     "qrcode.react": "^1.0.1",
-    "rc-picker": "^2.5.19",
     "react": "^17.0.2",
-    "react-animated-css": "^1.2.1",
     "react-datetime-picker": "^3.4.2",
     "react-dom": "^17.0.2",
     "react-geocode": "^0.2.3",
     "react-google-autocomplete": "^2.5.0",
     "react-google-maps": "^9.4.5",
     "react-icons": "^4.3.1",
-    "react-instantsearch": "^6.15.0",
     "react-loading": "^2.0.3",
     "react-redux": "^7.2.5",
     "react-ripples": "^2.2.1",
@@ -46,7 +39,6 @@
     "slick-carousel": "^1.8.1",
     "styled-components": "^5.3.1",
     "supercluster": "^7.1.4",
-    "swr": "^1.0.1",
     "use-supercluster": "^0.3.0",
     "uuid": "^8.3.2",
     "web-vitals": "^1.1.2"
