--- src/components/aboutPage/AboutPage.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const AboutPage = () => {
-  return <div>About</div>;
-};
-
-export default AboutPage;
