--- src/components/contactPage/ContactPage.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const ContactPage = () => {
-  return <div>Contact</div>;
-};
-
-export default ContactPage;
