+++ src/components/landingPage/Banner.jsx
@@ -3,21 +3,29 @@ import styled from 'styled-components';
 import { Link } from 'react-router-dom';
 import BannerImage from '../../images/homepage/sharefood-1.jpg';
 import HomepageBackground from '../common/HomepageBackground';
+import {
+  BannerContent,
+  BannerTitle,
+  Subtitle,
+  ButtonRow,
+  StartButton,
+  LearnMoreButton,
+} from '../common/Banner/CommonBanner';
 
 const Banner = () => {
   return (
     <BannerContainer>
       <HomepageBackground />
-      <BannerContent>
+      <StyledBannerContent>
         <BannerTitle>分享</BannerTitle>
-        <Description>
+        <Subtitle>
           透過交換平台分享剩食幫助別人與環境，一起把「剩食」變成「勝食」吧
-        </Description>
+        </Subtitle>
         <ButtonRow>
           <StartButton to="/login">開始使用</StartButton>
           <LearnMoreButton to="/search">了解更多</LearnMoreButton>
         </ButtonRow>
-      </BannerContent>
+      </StyledBannerContent>
       <BannerImg src={BannerImage} />
     </BannerContainer>
   );
@@ -31,7 +39,6 @@ const BannerContainer = styled.div`
   position: relative;
   padding: 5rem 2rem;
   background: linear-gradient(to right, #4ac29a, #bdfff3);
-  background: -webkit-linear-gradient(to left, #4ac29a, #bdfff3);
   backdrop-filter: blur(5px);
 
   @media screen and (min-width: 1500px) {
@@ -43,64 +50,11 @@ const BannerContainer = styled.div`
   }
 `;
 
-const BannerContent = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  max-width: 35vw;
+const StyledBannerContent = styled(BannerContent)`
   @media screen and (max-width: 600px) {
     max-width: 80vw;
     margin-bottom: 3rem;
   }
-`;
-const BannerTitle = styled.h2`
-  margin-bottom: 2rem;
-  font-family: 'cwTeXYen', sans-serif;
-  font-size: 56px;
-  font-weight: 600;
-  color: black;
-`;
-
-const Description = styled.p`
-  font-family: 'cwTeXYen', sans-serif;
-  font-size: 24px;
-  font-weight: 100;
-  margin-bottom: 2rem;
-  line-height: 2rem;
-  color: #0000009e;
-
-  @media screen and (min-width: 375px) {
-    max-width: 60vw;
-  }
-
-  @media screen and (min-width: 1500px) {
-    max-width: 20vw;
-  }
-`;
-
-const ButtonRow = styled.div`
-  display: flex;
-  flex-direction: row;
-`;
-
-const StyledLink = styled(Link)`
-  margin-right: 0.5rem;
-  padding: 0.5rem;
-  border-radius: 5px;
-  font-family: 'cwTeXYen', sans-serif;
-  font-size: 22px;
-`;
-
-const StartButton = styled(StyledLink)`
-  color: white;
-  background-color: #1e88e5;
-`;
-
-const LearnMoreButton = styled(StyledLink)`
-  background-color: white;
-  color: #52b788;
-  opacity: 0.8;
 `;
 
 const BannerImg = styled.img`
